requirements.txt
streamlit==1.27.0
Pillow==10.0.1
pytesseract==0.3.10
python-docx==0.8.11
numpy==1.26.0
tesseract-ocr
libtesseract-dev